@@ -5,8 +5,9 @@ author: izolate <izolate@gmail.com>
 homepage: https://github.com/izolate/front-matter
 documentation: https://github.com/izolate/front-matter/blob/master/README.md
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  # sdk: '>=2.0.0 <3.0.0'
+  sdk: ">=2.12.0 <4.0.0"
 dependencies:
-  yaml: ^2.1.15
+  yaml: ^3.1.1
 dev_dependencies:
-  test: ^1.6.2
+  test: ^1.21.4
